rename halotools module
